requirements.txt
# Core ML/AI
torch==2.9.0
transformers==4.57.1
sentence-transformers==5.1.2
faiss-cpu==1.12.0
huggingface-hub==0.36.0
numpy==2.2.6

# LLM APIs
anthropic==0.72.0
openai==2.6.1

# Audio processing
sounddevice==0.5.3
soundfile==0.13.1
librosa==0.11.0
vosk==0.3.45
webrtcvad==2.0.10
openai-whisper==20250625

# Web/API
fastapi==0.121.0
uvicorn==0.38.0
aiohttp==3.13.2
websockets==15.0.1
requests==2.32.5
Flask==3.1.2
Jinja2==3.1.6

# Data/Storage
chromadb==1.3.0
pandas==2.3.3

# System/Monitoring
psutil==7.1.3
prometheus-fastapi-instrumentator==7.1.0
paho-mqtt==2.1.0
inotify==0.2.10

# Message Bus
pyzmq==27.1.0

# Scientific computing
scipy==1.16.3
matplotlib==3.10.7

# Testing
pytest==8.4.2

# Utilities
pydantic==2.12.4
PyYAML==6.0.3
jsonschema==4.25.1
rank-bm25==0.2.2
rapidfuzz==3.14.0
schedule==1.2.2
tqdm==4.67.1
srt==3.5.3

# Browser automation
playwright==1.55.0

# System dependencies
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
pycparser==2.23
urllib3==2.5.0
